requirements.txt
asn1crypto==1.5.1
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==2.2.1
cryptography==44.0.1
filelock==3.17.0
idna==3.10
numpy==2.2.3
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
protobuf==5.29.3
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyrate-limiter==3.7.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
sec-edgar-downloader==5.0.3
six==1.17.0
snowflake-connector-python==3.13.2
snowflake-snowpark-python==1.28.0
sortedcontainers==2.4.0
soupsieve==2.6
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
